Add overrides for preformatting on a per sink basis.
Add compile flag lager_do_log_override_modules. This parameter
takes a proplists that maps sink -> module. The parse transform
uses the module for the specified sink when calling do_log and
do_log_unsafe. If there is no override specified then the default
lager behaviors are used.

The parametes for do_log and do_log_safe have not changed. THe return
value is now expected to be a fun/0 that will return the formatted
message.